pyproject.toml
[tool.poetry]
name = "django-equilibrium"
version = "0.1.0"
description = ""
authors = ["Xanya <progamerr99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
pillow = "^10.3.0"
django-image-cropping = "^1.7"
easy-thumbnails = "^2.8.5"
django-ckeditor = "^6.7.1"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
